restructure dfc code a bit
