+++ MPCAutofill/cardpicker/tests/test_update_dfcs.py
@@ -5,7 +5,7 @@ import requests
 from cardpicker.constants import DFC_URL, MELD_URL
 
 
-class TestScryfallAPI:
+class TestUpdateDFCs:
     # region tests
     def test_scryfall_queries(self) -> None:
         assert (response := requests.get(DFC_URL)).status_code == 200 and len(json.loads(response.content)["data"]) > 0
